requirements.txt
flask
opencv-python
pillow
ultralytics
numpy